framework: Allow specifying max wallpaper width
For some devices with very high-res screen (1920x1200), the wallpaper
width suggested by the launcher may be too large for the GPU and so
we need to limit it. (2048 for the above mentioned screen)

We don't need to limit the height since that's the height of the
screen.

Change-Id: I188ac585370fb705cfa50d5ad84786e6f2cbd016

Conflicts:
	core/res/res/values/config.xml
	core/res/res/values/public.xml
	services/java/com/android/server/WallpaperManagerService.java

Conflicts:
	core/res/res/values/config.xml
	core/res/res/values/symbols.xml